Corrected for ROSS rng calls type

+++ model/model.h
@@ -79,7 +79,7 @@ void lp_log(const char *,
 tw_lpid transition_select(tw_lp *,
 			  const struct transition *,
 			  tw_lpid,
-			  unsigned int *);
+			  long int *);
 int human_departure_events(struct state *, tw_lp *);
 
 #endif // MODEL_H
